requirements.txt
tk
pillow
opencv-python
pandas
scikit-image
scikit-learn